+++ tests/breadcrumb.rs
@@ -9,14 +9,14 @@
 mod util;
 
 use anyhow::Result;
-use sentry::{Breadcrumb, Event};
+use sentry::{Breadcrumb, Event, Options};
 use sentry_contrib_native as sentry;
 use serde_json::Value;
 
 #[tokio::test(flavor = "multi_thread")]
 async fn breadcrumb() -> Result<()> {
     util::events_success(
-        None,
+        Some(|options: &mut Options| options.set_max_breadcrumbs(4)),
         vec![
             (
                 || {
@@ -97,6 +97,34 @@ async fn breadcrumb() -> Result<()> {
                             .and_then(Value::as_object)
                             .and_then(|v| v.get("test data"))
                             .and_then(Value::as_str)
+                    );
+                },
+            ),
+            (
+                || {
+                    Breadcrumb::new(None, Some("test message 1".into())).add();
+                    Breadcrumb::new(None, Some("test message 2".into())).add();
+                    Event::new().capture()
+                },
+                |event| {
+                    assert_eq!("<unlabeled event>", event.title);
+                    let breadcrumbs = event
+                        .entries
+                        .get("breadcrumbs")
+                        .and_then(|v| v.get("values"))
+                        .and_then(Value::as_array)
+                        .unwrap();
+
+                    assert_eq!(breadcrumbs.len(), 4);
+
+                    let breadcrumb = breadcrumbs.get(3).and_then(Value::as_object).unwrap();
+                    assert_eq!(
+                        Some("default"),
+                        breadcrumb.get("type").and_then(Value::as_str)
+                    );
+                    assert_eq!(
+                        Some("test message 2"),
+                        breadcrumb.get("message").and_then(Value::as_str)
                     );
                 },
             ),